Create scan file service for tile
